refactor(open-mpm): split api/server.rs (3360 lines) into focused modules (closes #364)
Convert api/server.rs into an api/server/ module directory with one file per
logical concern, all under the 500-line cap:

- state.rs              AppState + task store + atomic persistence
- auth.rs               ApiConfig + bearer-token middleware
- routes.rs             router assembly + serve* bootstrap
- handlers.rs           task / health / docs core handlers
- projects.rs           project / session / agent listing
- project_registration.rs  project register + per-project config lookup
- ctrl_sessions.rs      CTRL session CRUD (om session ...)
- tm.rs                 tmux session management (/api/tm/*)
- events_sse.rs         SSE telemetry stream
- task_runner.rs        subprocess workflow execution + recap dispatch
- ui.rs                 embedded Vite bundle serving
- tests/{mod,ctrl_sessions,listing}.rs  test module split

Pure reorganization: no behavior change. The public interface
(serve_with_config, ApiConfig, AppState, TaskRequest, build_router*) is
re-exported from mod.rs, so consumers compile unchanged. All 1856 lib tests
plus integration tests pass.
